# cmd/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для библиотеки GOtus.",
        "title": "GOtus API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/reservation": {
            "post": {
                "description": "Добавляет новую запись бронирования книги",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Создать новую бронь",
                "parameters": [
                    {
                        "description": "Данные брони",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.CreateReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/reservation/{id}": {
            "get": {
                "description": "Возвращает одну запись бронирования по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Получить бронь по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gotus_internal_model_reservation.Reservation"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет запись бронирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Обновить бронь по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные брони",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.UpdateReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет запись бронирования по ID",
                "tags": [
                    "reservation"
                ],
                "summary": "Удалить бронь по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/reservations": {
            "get": {
                "description": "Возвращает все записи бронирования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservation"
                ],
                "summary": "Получить список всех броней",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gotus_internal_model_reservation.Reservation"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_reservation.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gotus_internal_model_reservation.Reservation": {
            "type": "object",
            "properties": {
                "BookInstanceID": {
                    "type": "integer"
                },
                "EndDate": {
                    "type": "string"
                },
                "Id": {
                    "type": "integer"
                },
                "ReservationStatusID": {
                    "type": "integer"
                },
                "StartDate": {
                    "type": "string"
                },
                "UserID": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_reservation.CreateReservationRequest": {
            "type": "object",
            "properties": {
                "book_instance_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_api_reservation.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_api_reservation.UpdateReservationRequest": {
            "type": "object",
            "properties": {
                "book_instance_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}